+++ Example/PinLayoutSample/UI/Examples/MultiRelativeView/MultiRelativeView.swift
@@ -20,21 +20,19 @@
 import UIKit
 import PinLayout
 
-class MultiRelativeView: BaseView {
-    fileprivate let contentView = UIView()
+class MultiRelativeView: UIView {
     fileprivate let view1 = BasicView(text: "Relative view 1 (width: 20%, height: 50%)", color: .lightGray)
     fileprivate let view2 = BasicView(text: "Relative view 2 (width: 20%, height: 50%)", color: .lightGray)
     fileprivate let view = BasicView(text: "View layouted using two relative views: \n  - right(of: view1, aligned: .top)\n  - left(of: view2, aligned: .bottom)",
                                      color: .pinLayoutColor)
     
-    override init() {
-        super.init()
+    init() {
+        super.init(frame: .zero)
+        backgroundColor = .white
     
-        addSubview(contentView)
-
-        contentView.addSubview(view1)
-        contentView.addSubview(view2)
-        contentView.addSubview(view)
+        addSubview(view1)
+        addSubview(view2)
+        addSubview(view)
     }
     
     required init?(coder aDecoder: NSCoder) {
@@ -44,11 +42,8 @@ class MultiRelativeView: BaseView {
     override func layoutSubviews() {
         super.layoutSubviews()
         
-        // Layout the contentView using the view's safeArea.
-        contentView.pin.all().margin(safeArea)
-        
-        view1.pin.top().left().width(20%).height(50%)
-        view2.pin.top().right().width(20%).height(50%)
+        view1.pin.top(pin.safeArea).left(pin.safeArea).width(20%).height(50%)
+        view2.pin.top(pin.safeArea).right(pin.safeArea).width(20%).height(50%)
         
         view.pin.right(of: view1, aligned: .top).left(of: view2, aligned: .bottom).marginHorizontal(10)
     }